refactor(testCases): 테스트 케이스 파일 testCases 디렉토리로 이동
